7.0_rl/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
python-dotenv==1.0.0
cassandra-driver==3.28.0
redis==4.6.0
dash==2.14.2
dash-bootstrap-components==1.5.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
gunicorn==21.2.0 